fix: resolve remaining audit items (R1/R2 doc-sync + M1/M3/M4/M5/M9 + LOWs)
R1: EXPERIMENTS.md §3 pins the n=150 artifact as the canonical quantization verdict and
    notes RESULTS.md may show a separate vram-run (report.py keeps newest per scope).
R2: EXPERIMENTS.md §4 marks the DeepSpeed 'nvcc absent' claim UNVERIFIED (no artifact;
    unsloth log shows CUDA Toolkit present).
M1: prepare_data default data_path -> raw_data/new_dialect (was a non-existent data/);
    old-dialect runbook note added.
M3: document the orphan '[ ]' strip in the cleaning-convention header.
M4: classifier class_weights tensor built as float32 (decoupled from fp16 activations).
M5: classifier.yaml comments eval_steps/save_steps as dead under strategy=epoch.
M9: grpo_trainer 'hm' aggregation now raises if normalize_rewards is off.
LOW: r_style uses a hard DO_TO_LABEL[...] lookup (surfaces unknown regions);
     epsilon_high>=epsilon DAPO guard; gguf n_gpu_layers default -1; serving.py
     latency guard uses 'is not None'.
AUDIT.md resolution table updated; M6/M7/B2 left by design (documented).

+++ src/ko_dialect/models/classifier.py
@@ -130,8 +130,10 @@ class TextCNNForSequenceClassification(PreTrainedModel):
         if labels is not None:
             weight = None
             if self.config.class_weights is not None:
+                # float32 for the weight tensor regardless of activation dtype (fp16);
+                # CrossEntropyLoss accepts a float32 weight with fp16 logits.
                 weight = torch.as_tensor(
-                    self.config.class_weights, dtype=logits.dtype, device=logits.device
+                    self.config.class_weights, dtype=torch.float32, device=logits.device
                 )
             loss = nn.CrossEntropyLoss(weight=weight)(logits, labels)
